split containers components from view compontes
